third_party/WebKit/LayoutTests/fast/hidpi/dpr-setting.html

<html>
<head>
<script>
    window.targetScaleFactor = 2;
</script>
<script src="resources/srcset-helper.js"></script>
<script src="../js/resources/js-test-pre.js"></script>
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
    }

    addEventListener("load", function() {
        shouldBeTrue('matchMedia("screen and (min-resolution: 2dppx)").matches');
    }, false);
</script>
</head>

<body id="body">
    <div>This test passes if the DPR is set to 2x</div>
</body>
</html>
