third_party/WebKit/LayoutTests/fast/hidpi/dpr-setting-expected.txt

PASS matchMedia("screen and (min-resolution: 2dppx)").matches is true
PASS successfullyParsed is true

TEST COMPLETE
This test passes if the DPR is set to 2x
